Revert "try and reenable qsort for 386 (again)"
This reverts commit 367f37eee70c0461f63b394cef49399a50130083.

+++ func_test.go
@@ -74,10 +74,10 @@ func Test_qsort(t *testing.T) {
 
 	data := []int{88, 56, 100, 2, 25}
 	sorted := []int{2, 25, 56, 88, 100}
-	compare := func(a, b *int) uintptr {
-		return uintptr(*a - *b)
+	compare := func(a, b *int) int32 {
+		return int32(*a - *b)
 	}
-	var qsort func(data []int, nitms uintptr, size uintptr, compar func(a, b *int) uintptr)
+	var qsort func(data []int, nitms uintptr, size uintptr, compar func(a, b *int) int32)
 	purego.RegisterLibFunc(&qsort, libc, "qsort")
 	qsort(data, uintptr(len(data)), unsafe.Sizeof(int(0)), compare)
 	for i := range data {
